move testcases to internal/testcases and update imports
