+++ internal/testcases/iddeclarations/shared/shared.go
