+++ internal/testcases/typeresolution/pkg1/pkg1.go
