+++ internal/testcases/typeresolution/pkg2/pkg2.go
